refactor: rename `process.rs` -> `shell.rs`
