+++ crates/z-shell/src/shell.rs
